[fix][chaejy]: api 분이
